Add files via upload
Lab Objective:
The goal of this lab is to enable you to acquire the skills and knowledge necessary to modify the native VLAN to a VLAN other than VLAN 1 and to disable unused switch ports, thereby enhancing switch security and preventing unauthorized network access.

Lab Purpose:
This lab focuses on securing a switch by mastering the process of altering the native VLAN and deactivating unused ports. These techniques are essential to protect the network from potential security breaches if an unauthorized device is connected to an active port.

In this lab, I configured a chosen interface to function in trunk mode and set VLAN 20 as the native VLAN for the trunk link, ensuring proper VLAN traffic management. Next, I disabled FastEthernet 10 through 15 switch ports, effectively shutting them down to prevent unauthorized access. Finally, I executed the necessary show commands to confirm that my configurations were correctly applied, providing evidence of the changes made to the switch.